Custom skeleton for my maven projects
Changed directory structure to be rid of the standard nested structure
of maven.

Changed final .jar name to "run.jar" so its execution or deployment
can easily be automated even if the version or ids change.

Automation scripts "run.sh" and "test.sh" can be bound to the respective
keys in your working environment, i.e. in VIM with ',r' and ',t'.